+++ tests/TenCenturiesServiceTests.swift
@@ -1,0 +1,8 @@
+import XCTest
+@testable import Macchiato
+
+class TenCenturiesServiceTests {
+    func testBaseUrlIsAUrl() {
+        XCTAssertNotNil(TenCenturies.baseURL)
+    }
+}